+++ CHANGELOG.md
@@ -36,7 +36,8 @@
 ### Bug修复
 * 【core   】     修复SimpleCache死锁问题（issue#I1HOKB@Gitee）
 * 【core   】     修复SemaphoreRunnable释放问题（issue#I1HLQQ@Gitee）
-* 【poi    】     修复Sax方式读取Excel行号错误问题（issue#882@Gitee）
+* 【poi    】     修复Sax方式读取Excel行号错误问题（issue#882@Github）
+* 【poi    】     修复Sax方式读取Excel日期类型数据03和07不一致问题（issue#I1HL1C@Gitee）
 
 -------------------------------------------------------------------------------------------------------------